+++ lamindb/dev/db/_staged.py
@@ -3,11 +3,12 @@ from typing import Any
 
 import lnschema_core as core
 
-from .. import get_name_suffix_from_filepath, track_usage
+from .. import get_name_suffix_from_filepath
 from .._insert import insert
 from ..file import store_file
 from ..object import infer_suffix, write_to_file
 from ._linkstaged import LinkStaged
+from ._track_usage import track_usage
 
 
 class Staged:
